cadl, use Java Integer for Cadl int32 or shorter (#1680)
* use java integer for cadl int32 or shorter

* update tests (with cadl-ranch updated)

* manual fix due to int type change

* include more cadl-ranch sources

+++ cadl-tests/src/main/java/com/cadl/optional/models/Optional.java
@@ -74,7 +74,7 @@ public final class Optional {
      * The int property.
      */
     @JsonProperty(value = "int")
-    private Long intProperty;
+    private Integer intProperty;
 
     /*
      * The long property.
@@ -279,7 +279,7 @@ public final class Optional {
      *
      * @return the intProperty value.
      */
-    public Long getIntProperty() {
+    public Integer getIntProperty() {
         return this.intProperty;
     }
 
@@ -289,7 +289,7 @@ public final class Optional {
      * @param intProperty the intProperty value to set.
      * @return the Optional object itself.
      */
-    public Optional setIntProperty(Long intProperty) {
+    public Optional setIntProperty(Integer intProperty) {
         this.intProperty = intProperty;
         return this;
     }
